Remove nose based test file, and move tests to requests.py
